[Fix #72] Rest Time 설정오류 수정
